requirements.txt
python-dotenv
langchain
langchain_community
pydantic
langchain-mcp-tools
mcp 
openai
langchain-openai
azure-devops
mcp_registry